+++ CHANGELOG.md
@@ -3,11 +3,11 @@
 All notable changes to the "todo-tools" extension will be documented in this file.
 
 ## [0.8.9]
+### add
+- recur(Mon,Tue,Wed)
+
 ### fix
 - recur(weekday) is recurring on the same day as done
-
-### todo:
-- recur(Mon,Tue,Wed)
 
 ## [0.8.8] - 2025-06-18
 ### add